hw7

+++ src/com/pb/kopiton/hw7/Tshirt.java
@@ -1,7 +1,7 @@
 package com.pb.kopiton.hw7;
 
 public class Tshirt extends Clothes implements ManClothes, WomenClothes {
-    public Tshirt(Size size, float price, String color) {
+    public Tshirt(Size size, int price, String color) {
         super(size, price, color);
     }
